requirements.txt
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
binaryornot==0.4.4
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==4.2.11
django-browser-reload==1.13.0
django-extensions==3.2.3
django-tailwind==3.6.0
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.51.0
future==1.0.0
gpiozero==2.0.1
idna==3.7
importlib-metadata==4.13.0
importlib-resources==5.13.0
ipykernel==6.29.4
ipython==8.18.1
iso8601==2.1.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
nest_asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pkg_resources==0.0.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py_ballisticcalc==2.0.1
py_ballisticcalc.exts==2.0.1
Pygments==2.17.2
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.32.3
rich==13.7.1
six==1.16.0
smbus==1.1.post2
spidev==3.6
sqlparse==0.5.0
stack-data==0.6.3
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.18.1
